refactor: Migrate private and session card components to Angular signals and introduce a new private module.
