+++ src/app/modules/private/private.page.ts
@@ -1,36 +1,6 @@
-import {
-  Component,
-  HostListener,
-  NgZone,
-  OnInit,
-  signal
-} from '@angular/core';
-import { Router } from '@angular/router';
-import { AlertController, MenuController, Platform } from '@ionic/angular';
-import { TranslateService } from '@ngx-translate/core';
-import { localKeys } from 'src/app/core/constants/localStorage.keys';
-import {
-  AuthService,
-  DbService,
-  HttpService,
-  LocalStorageService,
-  NetworkService,
-  UserService,
-  UtilService,
-} from 'src/app/core/services';
-import { ProfileService } from 'src/app/core/services/profile/profile.service';
-import { CommonRoutes } from 'src/global.routes';
-import { App, URLOpenListenerEvent } from '@capacitor/app';
-import { environment } from 'src/environments/environment';
-import { Location } from '@angular/common';
-import * as _ from 'lodash-es';
-import { PermissionService } from '../../core/services/permission/permission.service';
-import { permissionModule } from 'src/app/core/constants/permissionsConstant';
+import { Component, OnInit, computed } from '@angular/core';
 import { PAGE_IDS } from 'src/app/core/constants/page.ids';
-import {
-  FrontendChatLibraryService,
-  RocketChatApiService,
-} from 'sl-chat-library';
+import { PrivateService } from 'src/app/core/services/private.service';
 
 @Component({
   selector: 'app-private',
@@ -39,409 +9,18 @@ import {
   standalone: false
 })
 export class PrivatePage implements OnInit {
-  user = signal<any>(null);
   PAGE_IDS = PAGE_IDS;
-  public appPages = [
-    {
-      title: 'HOME',
-      action: 'home',
-      icon: 'home',
-      class: 'hide-on-small-screen',
-      url: CommonRoutes.TABS + '/' + CommonRoutes.HOME,
-      pageId: PAGE_IDS.home,
-    },
-    {
-      title: 'MENTORS',
-      action: 'mentor-directory',
-      icon: 'people',
-      class: 'hide-on-small-screen',
-      url: CommonRoutes.TABS + '/' + CommonRoutes.MENTOR_DIRECTORY,
-      pageId: PAGE_IDS.mentorDirectory,
-    },
-    {
-      title: 'REQUESTS',
-      action: 'requests',
-      icon: 'mail',
-      url: CommonRoutes.TABS + '/' + CommonRoutes.REQUESTS,
-      pageId: PAGE_IDS.requests,
-      badge: false,
-    },
-    {
-      title: 'MY_CONNECTIONS',
-      action: 'my-connections',
-      icon: 'people',
-      url: CommonRoutes.MY_CONNECTIONS,
-      pageId: PAGE_IDS.myConnections,
-    },
-    {
-      title: 'MESSAGES',
-      action: 'messages',
-      icon: 'chatbubbles',
-      url: CommonRoutes.MESSAGES,
-      badge: false,
-      pageId: PAGE_IDS.messages,
-    },
-    {
-      title: 'DASHBOARD',
-      action: 'dashboard',
-      icon: 'stats-chart',
-      class: 'hide-on-small-screen',
-      url: CommonRoutes.TABS + '/' + CommonRoutes.DASHBOARD,
-      pageId: PAGE_IDS.dashboard,
-    },
-    {
-      title: 'HELP_SIDEMENU',
-      action: 'help',
-      icon: 'help-circle',
-      url: CommonRoutes.HELP,
-      pageId: PAGE_IDS.help,
-    },
-    {
-      title: 'FAQ',
-      action: 'faq',
-      icon: 'alert-circle',
-      url: CommonRoutes.FAQ,
-      pageId: PAGE_IDS.faq,
-    },
-    {
-      title: 'USER_GUIDE_VIDEOS',
-      action: 'help videos',
-      icon: 'videocam',
-      url: CommonRoutes.HELP_VIDEOS,
-      pageId: PAGE_IDS.helpVideos,
-    },
-    {
-      title: 'LANGUAGE',
-      action: 'selectLanguage',
-      icon: 'language',
-      url: CommonRoutes.LANGUAGE,
-      pageId: PAGE_IDS.language,
-    },
-    {
-      title: 'BLOCKED_USERS',
-      action: 'blocked-users',
-      icon: 'ban',
-      url: CommonRoutes.BLOCKED_USERS,
-      pageId: PAGE_IDS.blockedUsers,
-    },
-    {
-      title: 'CHANGE_PASSWORD',
-      action: 'change-password',
-      icon: 'key',
-      url: CommonRoutes.CHANGE_PASSWORD,
-      pageId: PAGE_IDS.changePassword,
-    },
-    {
-      title: 'LOGIN_ACTIVITY',
-      action: 'login-activity',
-      icon: 'time',
-      url: CommonRoutes.LOGIN_ACTIVITY,
-      pageId: PAGE_IDS.loginActivity,
-    },
-  ];
 
-  adminPage = {
-    title: 'ADMIN_WORKSPACE',
-    action: 'admin',
-    icon: 'briefcase',
-    class: '',
-    url: CommonRoutes.ADMIN + '/' + CommonRoutes.ADMIN_DASHBOARD,
-    pageId: PAGE_IDS.adminWorkspace,
-  };
+  roleLabel = computed(() =>
+    this.privateService.isMentor() ? 'MENTOR' : 'MENTEE'
+  );
 
-  actionsArrays: any[] = permissionModule.MODULES;
-
-  isMentor = signal<boolean>(false);
-  showAlertBox = false;
-  userRoles = signal<any>(null);
-  userEventSubscription: any;
-  backButtonSubscription: any;
-  menuSubscription: any;
-  routerSubscription: any;
-  adminAccess = signal<boolean>(false);
-
-  isAuthBypassed = environment['isAuthBypassed'];
   constructor(
-    private translate: TranslateService,
-    private platform: Platform,
-    private localStorage: LocalStorageService,
-    public menuCtrl: MenuController,
-    private userService: UserService,
-    private utilService: UtilService,
-    private db: DbService,
-    private router: Router,
-    private network: NetworkService,
-    private authService: AuthService,
-    private profile: ProfileService,
-    private zone: NgZone,
-    private _location: Location,
-    private alert: AlertController,
-    private permissionService: PermissionService,
-    private chatService: FrontendChatLibraryService,
-    private rocketChatService: RocketChatApiService
+    public privateService: PrivateService
   ) { }
 
   async ngOnInit() {
-    await this.initializeApp();
-    if (this.isMentor()) {
-      const response = await this.profile.getRequestCount();
-      const { result = {} } = response || {};
-      const { sessionRequestCount = 0, connectionRequestCount = 0 } = result || {};
-      if (sessionRequestCount > 0 || connectionRequestCount > 0) {
-        const page = this.appPages.find(
-          (page: any) => page.pageId === PAGE_IDS.requests
-        );
-        if (page) {
-          page.badge = true;
-        }
-      }
-      this.updateBadgeFlag();
-    }
-    await this.rocketChatService.initializeWebSocketAndCheckUnread();
-    if (this.chatService.initialBadge) {
-      let page = this.appPages.find(
-        (page: any) => page.pageId == PAGE_IDS.messages
-      );
-      page.badge = this.chatService.initialBadge;
-    }
-    this.updateBadgeFlag();
-    this.chatService.showBadge.subscribe((resp: boolean) => {
-      let page = this.appPages.find(
-        (page: any) => page.pageId == PAGE_IDS.messages
-      );
-      page.badge = resp;
-      this.updateBadgeFlag();
-    });
-  }
-
-  updateBadgeFlag() {
-    const hasBadge = this.appPages.some(p => p.badge);
-    this.utilService.setHasBadge(hasBadge);
-  }
-
-  subscribeBackButton() {
-    this.backButtonSubscription =
-      this.platform.backButton.subscribeWithPriority(10, async () => {
-        if (this._location.isCurrentPathEqualTo('/tabs/home')) {
-          let texts: any;
-          this.translate
-            .get(['EXIT_CONFIRM_MESSAGE', 'CANCEL', 'CONFIRM'])
-            .subscribe((text) => {
-              texts = text;
-            });
-          const alert = await this.alert.create({
-            message: texts['EXIT_CONFIRM_MESSAGE'],
-            buttons: [
-              {
-                text: texts['CANCEL'],
-                role: 'cancel',
-                cssClass: 'alert-button-bg-white',
-                handler: () => { },
-              },
-              {
-                text: texts['CONFIRM'],
-                role: 'confirm',
-                cssClass: 'alert-button',
-                handler: () => {
-                  navigator['app'].exitApp();
-                },
-              },
-            ],
-          });
-          await alert.present();
-        } else {
-          this._location.back();
-        }
-      });
-  }
-
-  async initializeApp() {
-    await this.platform.ready();
-    this.network.netWorkCheck();
-    let theme: any = localStorage.getItem('theme');
-    if (theme) {
-      try {
-        theme = JSON.parse(theme);
-        document.documentElement.style.setProperty('--ion-color-primary', theme.primaryColor);
-        document.documentElement.style.setProperty('--ion-color-secondary', theme.secondaryColor);
-      } catch (error) {
-        console.error("Error parsing theme from localStorage:", error);
-      }
-    }
-    // this.profile.getTheme();
-
-    await new Promise<void>((resolve) => {
-      setTimeout(async () => {
-        this.languageSetting();
-        this.setHeader();
-        const userDetails = await this.localStorage.getLocalData(
-          localKeys.USER_DETAILS
-        );
-
-        if (userDetails) {
-          this.profile.getUserRole(userDetails);
-          this.adminAccess.set(userDetails.permissions
-            ? this.permissionService.hasAdminAcess(
-              this.actionsArrays,
-              userDetails?.permissions
-            )
-            : false);
-        }
-        await this.profile.getChatToken();
-        this.getUser();
-        resolve();
-      }, 0);
-    });
-
-    this.db.init();
-
-    await new Promise<void>((resolve) => {
-      setTimeout(async () => {
-        this.userRoles.set(await this.localStorage.getLocalData(
-          localKeys.USER_ROLES
-        ));
-        resolve();
-      }, 1000);
-    });
-    setTimeout(() => {
-      document
-        .querySelector('ion-menu')
-        ?.shadowRoot?.querySelector('.menu-inner')
-        ?.setAttribute('style', 'border-radius:8px 8px 0px 0px');
-    }, 2000);
-    this.userEventSubscription = this.userService.userEventEmitted$.subscribe(
-      (data) => {
-        if (data) {
-          this.isMentor.set(this.profile.isMentor);
-          this.user.set(data);
-          this.adminAccess.set(data.permissions
-            ? this.permissionService.hasAdminAcess(
-              this.actionsArrays,
-              data?.permissions
-            )
-            : false);
-        }
-      }
-    );
-    App.addListener('appUrlOpen', (event: URLOpenListenerEvent) => {
-      this.zone.run(() => {
-        const domain = window.location.origin;
-        const slug = event.url.split(domain).pop();
-        if (slug) {
-          this.router.navigateByUrl(slug);
-        }
-      });
-    });
-    this.subscribeBackButton();
-  }
-  setHeader() {
-    this.userService.getUserValue();
-  }
-  languageSetting() {
-    this.localStorage
-      .getLocalData(localKeys.SELECTED_LANGUAGE)
-      .then((data) => {
-        if (data) {
-          this.translate.use(data);
-        } else {
-          this.setLanguage('en');
-        }
-      })
-      .catch((error) => {
-        this.setLanguage('en');
-      });
-  }
-
-  setLanguage(lang) {
-    this.localStorage
-      .setLocalData(localKeys.SELECTED_LANGUAGE, lang)
-      .then((data) => {
-        this.translate.use(lang);
-      })
-      .catch((error) => {
-        this.translate.use(lang);
-      });
-  }
-
-  logout() {
-    let msg = {
-      header: 'LOGOUT',
-      message: 'LOGOUT_CONFIRM_MESSAGE',
-      cancel: 'CANCEL',
-      submit: 'LOGOUT',
-    };
-    this.utilService
-      .alertPopup(msg)
-      .then(async (data) => {
-        if (data) {
-          await this.localStorage.setLocalData(
-            localKeys.SELECTED_LANGUAGE,
-            'en'
-          );
-          this.translate.use('en');
-          await this.authService.logoutAccount();
-          this.menuCtrl.enable(false);
-        }
-      })
-      .catch((error) => { });
-  }
-  goToProfilePage() {
-    this.menuCtrl.toggle();
-    this.router.navigate([`${CommonRoutes.TABS}/${CommonRoutes.PROFILE}`]);
-  }
-
-  async menuItemAction(menu) {
-    this.router.navigate([menu.url]);
-  }
-
-  ngOnDestroy(): void {
-    if (this.userEventSubscription) {
-      this.userEventSubscription.unsubscribe();
-    }
-    if (this.backButtonSubscription) {
-      this.backButtonSubscription.unsubscribe();
-    }
-    if (this.menuSubscription) {
-      this.menuSubscription.unsubscribe();
-    }
-    if (this.routerSubscription) {
-      this.routerSubscription.unsubscribe();
-    }
-  }
-
-  getUser() {
-    let theme: any = localStorage.getItem('theme');
-    if (theme) {
-      try {
-        theme = JSON.parse(theme);
-        document.documentElement.style.setProperty('--ion-color-primary', theme.primaryColor);
-        document.documentElement.style.setProperty('--ion-color-secondary', theme.secondaryColor);
-      } catch (error) {
-        console.error("Error parsing theme from localStorage:", error);
-      }
-    }
-    this.profile.getProfileDetailsFromAPI().then(profileDetails => {
-      if (profileDetails?.organizations && profileDetails?.organizations.length == 1) {
-        this.authService.setUserInLocal(profileDetails);
-      } else if (profileDetails?.organizations && profileDetails?.organizations.length > 1) {
-        // this.showOrganizationModal(profileDetails?.result?.user?.organizations);
-      }
-      this.adminAccess.set(profileDetails?.permissions ? this.permissionService.hasAdminAcess(this.actionsArrays, profileDetails?.permissions) : false);
-      this.user.set(profileDetails);
-      // !environment['isAuthBypassed'] && 
-      // !environment['isAuthBypassed'] && 
-      if (profileDetails?.profile_mandatory_fields && profileDetails?.profile_mandatory_fields.length > 0 || !profileDetails?.about) {
-        this.router.navigate([`/${CommonRoutes.EDIT_PROFILE}`], { replaceUrl: true, queryParams: { redirectUrl: '/tabs/home' } });
-      }
-      this.isMentor.set(this.profile.isMentor);
-    })
-  }
-
-  async viewRoles() {
-    const userRoles = await this.localStorage.getLocalData(
-      localKeys.USER_ROLES
-    );
-    this.profile.viewRolesModal(userRoles);
+    await this.privateService.initializeApp();
   }
 
   isCustomIcon(icon: string | undefined): boolean {
