+++ src/app/shared/components/session-card/session-card.component.html
@@ -1,45 +1,45 @@
-<ion-card (click)="onCardClick(data)" class="session-card card-shadow-bottom">
-  @if (data?.status?.value == 'LIVE') {
+<ion-card (click)="onCardClick(data())" class="session-card card-shadow-bottom">
+  @if (data()?.status?.value == 'LIVE') {
   <span class="badge-container status-badge d-flex flex-justify-center" [ngClass]="{
         'live-tag':
-          (data?.status?.value == 'LIVE' && data?.is_assigned) ||
-          data?.enrolled_type === 'INVITED'
+          (data()?.status?.value == 'LIVE' && data()?.is_assigned) ||
+          data()?.enrolled_type === 'INVITED'
       }">
-    {{ data?.status?.value | titlecase }}
+    {{ data()?.status?.value | titlecase }}
   </span>
   }
 
-  @if (data?.is_assigned) {
+  @if (data()?.is_assigned) {
   <span class="badge-container assigned-tag d-flex flex-justify-center">
     <ion-icon class="assigned-tag-icon" name="person-circle-sharp"></ion-icon>
     {{ "ASSIGNED" | translate }}
   </span>
   }
 
-  @if (data?.enrolled_type === 'INVITED' || data?.enrolment_type === 'INVITED') {
+  @if (data()?.enrolled_type === 'INVITED' || data()?.enrolment_type === 'INVITED') {
   <span class="badge-container assigned-tag d-flex flex-justify-center">
     <ion-icon class="assigned-tag-icon" name="person-circle-sharp"></ion-icon>
     {{ "INVITED" | translate }}
   </span>
   }
 
-  @if (showBanner) {
-  <img [src]="data?.image[0] || 'assets/default-session-img/session-img.png'" alt="Banner Image" class="banner-img" />
+  @if (showBanner()) {
+  <img [src]="data()?.image[0] || 'assets/default-session-img/session-img.png'" alt="Banner Image" class="banner-img" />
   }
 
   <ion-card-content>
     <ion-grid class="ion-no-padding">
       <ion-row>
         <div class="session-image">
-          <img class="thumbnail-image" [src]="data?.image[0] || 'assets/default-session-img/session-img.png'" />
+          <img class="thumbnail-image" [src]="data()?.image[0] || 'assets/default-session-img/session-img.png'" />
         </div>
         <ion-col padding size-xs="6" size-sm="8" size-md="6" size-lg="8">
           <div class="header">
-            <h5 class="text-align">{{ data?.name || data?.title }}</h5>
+            <h5 class="text-align">{{ data()?.name || data()?.title }}</h5>
             <div class="margin-btm-10">
-              @if (data?.description) {
+              @if (data()?.description) {
               <p class="text-align">
-                {{ data?.mentor_name }}
+                {{ data()?.mentor_name }}
               </p>
               }
             </div>
@@ -52,7 +52,7 @@
           <ion-icon name="calendar-number-outline"></ion-icon>
           <div>
 
-            @if (startDate() && data?.status?.value == 'PUBLISHED') {
+            @if (startDate() && data()?.status?.value == 'PUBLISHED') {
             <ion-label class="platform-date-label">
               {{ "STARTS_ON" | translate }}
               {{ startDate() | date : "dd/MM/yyyy" }} {{ "AT" | translate }}
@@ -60,7 +60,7 @@
             </ion-label>
             }
 
-            @if (startDate() && data?.status?.value == 'LIVE') {
+            @if (startDate() && data()?.status?.value == 'LIVE') {
             <ion-label class="platform-date-label">
               {{ "STARTED_ON" | translate }}
               {{ startDate() | date : "dd/MM/yyyy" }} {{ "AT" | translate }}
@@ -68,7 +68,7 @@
             </ion-label>
             }
 
-            @if (endDate() && data?.status?.value == 'COMPLETED') {
+            @if (endDate() && data()?.status?.value == 'COMPLETED') {
             <ion-label class="platform-date-label">
               {{ "COMPLETED_ON" | translate }}
               {{ endDate() | date : "dd/MM/yyyy" }} {{ "AT" | translate }}
@@ -78,29 +78,29 @@
           </div>
         </ion-row>
 
-        @if (!(meetingPlatform?.platform == 'OFF' && isCreator)) {
+        @if (!(meetingPlatform()?.platform == 'OFF' && isCreator())) {
         <ion-row class="d-flex flex-justify-center meeting-platform">
           <ion-icon name="videocam"></ion-icon>
           <div class="join-platform-container">
             <p class="platform-date-label">
               {{ "MEETING_PLATFORM" | translate }} :
-              <b>{{ meetingPlatform?.platform }}</b>
+              <b>{{ meetingPlatform()?.platform }}</b>
             </p>
           </div>
         </ion-row>
         }
 
         @if (
-        meetingPlatform?.platform == 'OFF' &&
-        isCreator &&
-        data.status.value != 'COMPLETED'
+        meetingPlatform()?.platform == 'OFF' &&
+        isCreator() &&
+        data()?.status?.value != 'COMPLETED'
         ) {
         <ion-row class="d-flex flex-align-items-center meeting-platform">
           <ion-icon name="videocam"></ion-icon>
           <div>
             <p class="platform-date-label">
               <b class="add-meeting-link"
-                (click)="clickOnAddMeetingLink(data); $event.stopPropagation()">{{"SELECT_MEETING_PLATFORM" |
+                (click)="clickOnAddMeetingLink(data()); $event.stopPropagation()">{{"SELECT_MEETING_PLATFORM" |
                 translate}}</b>
             </p>
           </div>
@@ -109,13 +109,13 @@
       </ion-col>
 
       <ion-col class="d-flex flex-justify-flex-end flex-align-items-center col-float" size="4">
-        @if (data?.status?.value != 'COMPLETED') {
+        @if (data()?.status?.value != 'COMPLETED') {
         <ion-button class="button" [disabled]="
-              !buttonConfig?.isEnabled && buttonConfig?.label != 'ENROLL'
+              !buttonConfig()?.isEnabled && buttonConfig()?.label != 'ENROLL'
             " (click)="
-              $event.stopPropagation(); onButtonClick(data, buttonConfig?.type)
+              $event.stopPropagation(); onButtonClick(data(), buttonConfig()?.type)
             " color="primary">
-          {{ buttonConfig?.label | translate }}
+          {{ buttonConfig()?.label | translate }}
         </ion-button>
         }
       </ion-col>
